Move tests for Study.optimize to test_optimize.py
